requirements.txt
fastapi
sqlalchemy
uvicorn
PyJWT
pydantic_settings
python-multipart
passlib[bcrypt]
pydantic[email]
pytest
asyncpg
